Update
Added admin privlages to delete decks.
Added a button on public decks that only admins can see to delete the deck.
Added proper permissions so users can' delete others decks.
Cleaned up the DBFunction file and sorted the functions by their category. (Deck, Card, Category, Users, Permissions)

+++ controllers/deleteDeck.php
@@ -1,6 +1,4 @@
 <?php
-
-
 	session_start();
 
 	include('../models/DBConnect.php');
@@ -15,9 +13,11 @@
 		$deckID = $_GET['DeckID'];
 	}
 
-	DeleteDeck($deckID);
-	$_SESSION['editDecksMessage'] = 'Deck successfully deleted.';
+	if(CheckDeleteDeckPermissions($deckID)){
+		DeleteDeck($deckID);
+		$_SESSION['editDecksMessage'] = 'Deck successfully deleted.';		
+	}
+
 	header("Location: ../views/myDecks.php");
-	
 
 ?>
